+++ src/m365/outlook/commands/mail/mail-send.ts
@@ -1,9 +1,13 @@
+import * as fs from 'fs';
+import * as path from 'path';
+import { AxiosRequestConfig } from 'axios';
 import auth, { Auth } from '../../../../Auth';
 import { Logger } from '../../../../cli/Logger';
 import GlobalOptions from '../../../../GlobalOptions';
 import request from '../../../../request';
 import GraphCommand from '../../../base/GraphCommand';
 import commands from '../../commands';
+import { validation } from '../../../../utils/validation';
 
 interface CommandArgs {
   options: Options;
@@ -19,7 +23,8 @@ interface Options extends GlobalOptions {
   bodyContents: string;
   bodyContentType?: string;
   importance?: string;
-  saveToSentItems?: string;
+  attachment?: string | string[];
+  saveToSentItems?: boolean;
 }
 
 class OutlookMailSendCommand extends GraphCommand {
@@ -52,7 +57,8 @@ class OutlookMailSendCommand extends GraphCommand {
         saveToSentItems: args.options.saveToSentItems,
         importance: args.options.importance,
         mailbox: typeof args.options.mailbox !== 'undefined',
-        sender: typeof args.options.sender !== 'undefined'
+        sender: typeof args.options.sender !== 'undefined',
+        attachment: typeof args.options.attachment !== 'undefined'
       });
     });
   }
@@ -89,6 +95,9 @@ class OutlookMailSendCommand extends GraphCommand {
         autocomplete: ['low', 'normal', 'high']
       },
       {
+        option: '--attachment [attachment]'
+      },
+      {
         option: '--saveToSentItems [saveToSentItems]'
       }
     );
@@ -103,14 +112,32 @@ class OutlookMailSendCommand extends GraphCommand {
           return `${args.options.bodyContentType} is not a valid value for the bodyContentType option. Allowed values are Text|HTML`;
         }
 
-        if (args.options.saveToSentItems &&
-          args.options.saveToSentItems !== 'true' &&
-          args.options.saveToSentItems !== 'false') {
+        if (args.options.saveToSentItems && !validation.isValidBoolean(args.options.saveToSentItems as any)) {
           return `${args.options.saveToSentItems} is not a valid value for the saveToSentItems option. Allowed values are true|false`;
         }
 
         if (args.options.importance && ['low', 'normal', 'high'].indexOf(args.options.importance) === -1) {
           return `'${args.options.importance}' is not a valid value for the importance option. Allowed values are low|normal|high`;
+        }
+
+        if (args.options.attachment) {
+          const attachments: string[] = typeof args.options.attachment === 'string' ? [args.options.attachment] : args.options.attachment;
+
+          for (const attachment of attachments) {
+            if (!fs.existsSync(attachment)) {
+              return `File with path '${attachment}' was not found.`;
+            }
+
+            if (!fs.lstatSync(attachment).isFile()) {
+              return `'${attachment}' is not a file.`;
+            }
+          }
+
+          const requestBody = this.getRequestBody(args.options);
+          // The max body size of the request is 4 194 304 chars before getting a 413 response
+          if (JSON.stringify(requestBody).length > 4_194_304) {
+            return 'Exceeded the max total size of attachments which is 3MB.';
+          }
         }
 
         return true;
@@ -124,38 +151,17 @@ class OutlookMailSendCommand extends GraphCommand {
       if (isAppOnlyAuth === true && !args.options.sender) {
         throw `Specify a upn or user id in the 'sender' option when using app only authentication.`;
       }
-
-      const requestOptions: any = {
+  
+      const requestOptions: AxiosRequestConfig = {
         url: `${this.resource}/v1.0/${args.options.sender ? 'users/' + encodeURIComponent(args.options.sender) : 'me'}/sendMail`,
         headers: {
           accept: 'application/json;odata.metadata=none',
           'content-type': 'application/json'
         },
         responseType: 'json',
-        data: {
-          message: {
-            subject: args.options.subject,
-            body: {
-              contentType: args.options.bodyContentType || 'Text',
-              content: args.options.bodyContents
-            },
-            toRecipients: this.mapEmailAddressesToRecipients(args.options.to.split(',')),
-            ccRecipients: this.mapEmailAddressesToRecipients(args.options.cc?.split(',')),
-            bccRecipients: this.mapEmailAddressesToRecipients(args.options.bcc?.split(',')),
-            importance: args.options.importance
-          },
-          saveToSentItems: args.options.saveToSentItems
-        }
+        data: this.getRequestBody(args.options)
       };
-
-      if (args.options.mailbox) {
-        requestOptions.data.message.from = {
-          emailAddress: {
-            address: args.options.mailbox
-          }
-        };
-      }
-
+  
       await request.post(requestOptions);
     }
     catch (err: any) {
@@ -163,16 +169,42 @@ class OutlookMailSendCommand extends GraphCommand {
     }
   }
 
-  private mapEmailAddressesToRecipients(emailAddresses: string[] | undefined): { emailAddress: { address: string }; }[] | undefined {
-    if (!emailAddresses) {
-      return emailAddresses;
+  private mapEmailAddressToRecipient(email: string | undefined): { emailAddress: { address: string }; } | undefined {
+    if (!email) {
+      return undefined;
     }
 
-    return emailAddresses.map(email => ({
+    return {
       emailAddress: {
         address: email.trim()
       }
+    };
+  }
+
+  private getRequestBody(options: Options): { message: any, saveToSentItems?: boolean } {
+    const attachments = typeof options.attachment === 'string' ? [options.attachment] : options.attachment;
+    const attachmentContents = attachments?.map(a => ({
+      '@odata.type': '#microsoft.graph.fileAttachment',
+      name: path.basename(a),
+      contentBytes: fs.readFileSync(a, { encoding: 'base64' })
     }));
+
+    return ({
+      message: {
+        subject: options.subject,
+        body: {
+          contentType: options.bodyContentType || 'Text',
+          content: options.bodyContents
+        },
+        from: this.mapEmailAddressToRecipient(options.mailbox),
+        toRecipients: options.to.split(',').map(mail => this.mapEmailAddressToRecipient(mail)),
+        ccRecipients: options.cc?.split(',').map(mail => this.mapEmailAddressToRecipient(mail)),
+        bccRecipients: options.bcc?.split(',').map(mail => this.mapEmailAddressToRecipient(mail)),
+        importance: options.importance,
+        attachments: attachmentContents
+      },
+      saveToSentItems: options.saveToSentItems
+    });
   }
 }